split widgets if necessary, update exports
