+++ src/browser.ts
@@ -1,4 +1,5 @@
 export * as Models from './models/index.js';
-export * as Views from './views/index.js';
 export * as OpsTypes from './ops-types.js';
+export * from './views/index.js';
 export * from './utils/index.js';
+export * from './core/index.js';
