+++ src/controllers/cli.ts
@@ -1,10 +1,10 @@
 /* env: node */
-import { Models, Controllers } from '@tinystacks/ops-core';
+import { Controllers, Provider } from '@tinystacks/ops-core';
 import { Cli as CliModel } from '../models/cli.js';
-import { CliEnvironmentProvider } from '../models/cli-environment-provider.js';
+import { CliEnvironmentProvider } from '../core/cli-environment-provider.js';
+import { Cli as CliProps } from '../ops-types.js';
 
 import WidgetController = Controllers.Widget;
-import Provider = Models.Provider;
 
 type CliOverrides = {
   clear?: boolean,
@@ -12,6 +12,11 @@ type CliOverrides = {
 }
 
 export class Cli extends CliModel implements WidgetController {
+  
+  static fromJson (object: CliProps): Cli {
+    return new Cli(object);
+  }
+
   async getData (providers?: Provider[], overrides?: CliOverrides): Promise<void> {
     // This is imported like this to avoid webpack throwing up over node dependencies when this widget is used
     //  in the frontend next app
@@ -30,7 +35,7 @@ export class Cli extends CliModel implements WidgetController {
       // if an override told it to run
       || (overrides && overrides.run === true);
 
-    const commands = [];
+    const commands: string[] = [];
     let allEnvVars = {};
     for (const envProvider of (providers || []) .filter(p => 'getCliEnvironment' in p)) {
       const cliEnv = await (envProvider as unknown as CliEnvironmentProvider).getCliEnvironment();
