+++ src/core/credentials-provider.ts
