+++ src/index.ts
@@ -3,3 +3,4 @@ export * as Models from './models/index.js';
 export * as Views from './views/index.js';
 export * as OpsTypes from './ops-types.js';
 export * from './utils/index.js';
+export * from './core/index.js';
