+++ src/utils/index.ts
@@ -1,1 +1,2 @@
+export * from './do-nothing.js';
 export * from './find-provider.js';
