+++ test/mocks/MockProvider.ts
@@ -1,6 +1,6 @@
-import { BaseProvider } from "@tinystacks/ops-core";
+import { Provider } from "@tinystacks/ops-core";
 
-class MockProvider extends BaseProvider {
+class MockProvider extends Provider {
   static type = 'MockProvider';
 }
